Refactoring lockin binaries
